+++ src/WireMock.Net/Transformers/Handlebars/HandlebarsJsonPath.cs
@@ -7,7 +7,7 @@ using System.Linq;
 using WireMock.Util;
 using WireMock.Validation;
 
-namespace WireMock.Transformers
+namespace WireMock.Transformers.Handlebars
 {
     internal static class HandlebarsJsonPath
     {
