+++ src/WireMock.Net/Transformers/Handlebars/HandlebarsRegex.cs
@@ -5,7 +5,7 @@ using HandlebarsDotNet;
 using WireMock.Util;
 using WireMock.Validation;
 
-namespace WireMock.Transformers
+namespace WireMock.Transformers.Handlebars
 {
     internal static class HandlebarsRegex
     {
