+++ src/WireMock.Net/Transformers/Handlebars/IHandlebarsContext.cs
@@ -1,0 +1,9 @@
+﻿using HandlebarsDotNet;
+
+namespace WireMock.Transformers.Handlebars
+{
+    interface IHandlebarsContext : ITransformerContext
+    {
+        IHandlebars Handlebars { get; set; }
+    }
+}
