--- src/WireMock.Net/Transformers/IHandlebarsContext.cs
@@ -1,9 +1,0 @@
-﻿using HandlebarsDotNet;
-
-namespace WireMock.Transformers
-{
-    interface IHandlebarsContext : ITransformerContext
-    {
-        IHandlebars Handlebars { get; set; }
-    }
-}
