+++ src/WireMock.Net/Transformers/Scriban/ScribanContext.cs
@@ -3,7 +3,7 @@ using Scriban;
 using WireMock.Handlers;
 using WireMock.Types;
 
-namespace WireMock.Transformers
+namespace WireMock.Transformers.Scriban
 {
     internal class ScribanContext : ITransformerContext
     {
