+++ src/WireMock.Net/WireMock.Net.csproj
@@ -79,7 +79,7 @@
 
     <PackageReference Include="Microsoft.AspNet.WebApi.OwinSelfHost" Version="5.2.6" />
     <PackageReference Include="System.ValueTuple" Version="4.5.0" />
-    <PackageReference Include="Scriban" Version="2.1.4" />
+    <PackageReference Include="Scriban.Signed" Version="2.1.4" />
   </ItemGroup>
 
   <ItemGroup Condition=" '$(TargetFramework)' == 'net46' ">
@@ -89,11 +89,11 @@
     <PackageReference Include="Microsoft.Owin.Hosting" Version="4.0.0" />
     <PackageReference Include="System.Net.Http" Version="4.3.3" />
     <PackageReference Include="System.ValueTuple" Version="4.5.0" />
-    <PackageReference Include="Scriban" Version="2.1.4" />
+    <PackageReference Include="Scriban.Signed" Version="2.1.4" />
   </ItemGroup>
 
   <ItemGroup Condition=" '$(TargetFramework)' == 'net461' ">
-    <PackageReference Include="Scriban" Version="2.1.4" />
+    <PackageReference Include="Scriban.Signed" Version="2.1.4" />
     <PackageReference Include="Microsoft.AspNetCore" Version="2.2.0" />
 
     <!-- https://github.com/WireMock-Net/WireMock.Net/issues/507 -->
@@ -106,11 +106,11 @@
     <PackageReference Include="System.Xml.XmlDocument" Version="4.3.0" />
     <PackageReference Include="System.Xml.XPath.XmlDocument" Version="4.3.0" />
     <PackageReference Include="System.ValueTuple" Version="4.5.0" />
-    <PackageReference Include="Scriban" Version="2.1.4" />
+    <PackageReference Include="Scriban.Signed" Version="2.1.4" />
   </ItemGroup>
 
   <ItemGroup Condition=" '$(TargetFramework)' == 'netstandard2.0' or '$(TargetFramework)' == 'netstandard2.1' ">
-    <PackageReference Include="Scriban" Version="3.3.2" />
+    <PackageReference Include="Scriban.Signed" Version="3.3.2" />
     <PackageReference Include="Microsoft.AspNetCore" Version="2.2.0" />
 
     <!-- https://github.com/WireMock-Net/WireMock.Net/issues/507 -->
@@ -119,7 +119,7 @@
 
   <ItemGroup Condition=" '$(TargetFramework)' == 'netcoreapp3.1' or '$(TargetFramework)' == 'net5.0'">
     <FrameworkReference Include="Microsoft.AspNetCore.App" />
-    <PackageReference Include="Scriban" Version="3.3.2" />
+    <PackageReference Include="Scriban.Signed" Version="3.3.2" />
   </ItemGroup>
 
   <ItemGroup>
